Updated the particle tree producer
